+++ tests/ui/const_value_deep_nesting.rs
@@ -1,0 +1,11 @@
+use batch_impl::batch_trait;
+
+// A constant value nested beyond 128 levels must error out (check_value_refs
+// guard) instead of overflowing the stack.
+batch_trait! {
+    @a = [[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]];
+    A: TraitA;
+}
+
+trait TraitA {}
+fn main() {}